refactor: move polling message stream to own file
